Add option to push or replace path in pushPath()
And also store pathname and uri in history state so that we can easily fall back to window.location.pathname if event.state is null. This catches the case when shiny::updateQueryString() is used to update the query string because the JS for that function sets the history stack item state to `null`.